+++ Source/Kernel/Drivers/VGA/vga.h
@@ -1,6 +1,8 @@
 #ifndef VGA_H
 #define VGA_H
 
+#include <stdint.h>
+
 #define VGA_BASE_ADDR 0xB8000
 #define VGA_SCREEN_COLUMNS 80
 #define VGA_SCREEN_ROWS 25
@@ -27,8 +29,8 @@
 // Position on screen
 typedef struct screen_pos
 {
-    int x;
-    int y;
+    uint16_t x;
+    uint16_t y;
 } screen_pos;
 
 // Color with blinking bit
@@ -63,7 +65,7 @@ typedef struct vga_character
 // Screen
 typedef union screen
 {
-    unsigned short value;
+    uint16_t value;
     vga_character c;
 } screen;
 
@@ -72,19 +74,19 @@ void vga_printchar(char c);
 void vga_printchar_color(char c, vga_color * color);
 void vga_printstring(char * str);
 void vga_printstring_color(char * str, vga_color * color);
-void vga_set_char(int x, int y, char c);
+void vga_set_char(uint16_t x, uint16_t y, char c);
 void vga_set_char_struct(screen_pos spos, char c);
-char vga_get_char(int x, int y);
+char vga_get_char(uint16_t x, uint16_t y);
 char vga_get_char_struct(screen_pos spos);
-void vga_set_color(int x, int y, vga_color col);
+void vga_set_color(uint16_t x, uint16_t y, vga_color col);
 void vga_set_color_struct(screen_pos spos, vga_color col);
-union vga_color vga_get_color(int x, int y);
+union vga_color vga_get_color(uint16_t x, uint16_t y);
 union vga_color vga_get_color_struct(screen_pos spos);
-void vga_set_character(int x, int y, vga_character c);
+void vga_set_character(uint16_t x, uint16_t y, vga_character c);
 void vga_set_character_struct(screen_pos spos, vga_character c);
-vga_character vga_get_character(int x, int y);
+vga_character vga_get_character(uint16_t x, uint16_t y);
 vga_character vga_get_character_struct(screen_pos spos);
-void vga_set_cursor_pos(int x, int y);
+void vga_set_cursor_pos(uint16_t x, uint16_t y);
 void vga_set_cursor_pos_struct(screen_pos spos);
 screen_pos vga_get_cursor_pos();
 void vga_clear_screen();
